Fix data race in AddTime
There was a data race in `Tachymeter.AddTime`.
`m.Count` was atomically incremented, but accessing `m.Times` slice
wasn't. As such, calling `AddTime()` from multiple goroutines triggered
a data race. The higher the number of goroutines and/or the smaller
the value of `m.Size`, the higher the probability of a data race
occurring.

Fixes #43

Add -race in travis-ci script